feat: move preferences and mic to single layout
