+++ src/frontend/src/routes/(single)/mission-control/+page.svelte
