+++ src/frontend/src/routes/(single)/preferences/+page.svelte
